refactor: 「FirebaseException発生時に表示するメッセージ」を作成する関数を別ファイルへ分離
